Refactored RelationalAlgebraStringExpression and Column to move them to abstract base class instead of pandas
