+++ neurolang/utils/relational_algebra_set/__init__.py
@@ -1,19 +1,17 @@
 import os
-from .pandas import RelationalAlgebraColumnInt, RelationalAlgebraColumnStr
+from .abstract import RelationalAlgebraColumnInt, RelationalAlgebraColumnStr, RelationalAlgebraStringExpression
 
 if os.getenv("NEURO_RAS_BACKEND", "pandas") == "sql":
     from .sql import (
         NamedRelationalAlgebraFrozenSet,
         RelationalAlgebraFrozenSet,
         RelationalAlgebraSet,
-        RelationalAlgebraStringExpression,
     )
 else:
     from .pandas import (
         NamedRelationalAlgebraFrozenSet,
         RelationalAlgebraFrozenSet,
         RelationalAlgebraSet,
-        RelationalAlgebraStringExpression,
     )
