+++ neurolang/utils/relational_algebra_set/abstract.py
@@ -2,6 +2,23 @@ from abc import abstractmethod, abstractproperty
 from collections.abc import MutableSet, Set
 
 
+class RelationalAlgebraStringExpression(str):
+    def __repr__(self):
+        return "{}{{ {} }}".format(self.__class__.__name__, super().__repr__())
+
+
+class RelationalAlgebraColumn:
+    pass
+
+
+class RelationalAlgebraColumnInt(int, RelationalAlgebraColumn):
+    pass
+
+
+class RelationalAlgebraColumnStr(str, RelationalAlgebraColumn):
+    pass
+
+
 class RelationalAlgebraFrozenSet(Set):
     def __init__(self, columns):
         raise NotImplementedError()
@@ -23,16 +40,10 @@ class RelationalAlgebraFrozenSet(Set):
         pass
 
     def is_dum(self):
-        return (
-            self.arity == 0 and
-            self.is_empty()
-        )
+        return self.arity == 0 and self.is_empty()
 
     def is_dee(self):
-        return (
-            self.arity == 0 and
-            not self.is_empty()
-        )
+        return self.arity == 0 and not self.is_empty()
 
     @abstractmethod
     def __contains__(self, element):
@@ -206,7 +217,6 @@ class NamedRelationalAlgebraFrozenSet(RelationalAlgebraFrozenSet):
 
 
 class RelationalAlgebraSet(RelationalAlgebraFrozenSet, MutableSet):
-
     @abstractmethod
     def add(self, value):
         pass
